+++ CHANGELOG.md
@@ -16,6 +16,10 @@ Fixes
 - Fixed undrawn elements not updating their area correctly even if their children are drawn or accessed
 - Fixed rare unbound recursion issue with auto-hiding scroll bars and elements that just about don't fit the panel
 
+### MLEM.Data
+Additions
+- Added the TypeHelper utility class
+
 ## 8.0.0
 
 ### MLEM